+++ build_runner_core/lib/src/generate/build_definition.dart
@@ -210,8 +210,8 @@ class _Loader {
           () => _updateAssetGraph(assetGraph, assetTracker, _buildPhases,
               inputSources, cacheDirSources, internalSources));
 
-      buildScriptUpdates = await BuildScriptUpdates.create(
-          _environment.reader, _options.packageGraph, assetGraph);
+      buildScriptUpdates = await BuildScriptUpdates.create(_environment.reader,
+          _options.packageGraph, assetGraph, _options.skipBuildScriptCheck);
       if (!_options.skipBuildScriptCheck &&
           buildScriptUpdates.hasBeenUpdated(updates.keys.toSet())) {
         _logger.warning('Invalidating asset graph due to build script update!');
@@ -240,7 +240,10 @@ class _Loader {
           throw CannotBuildException();
         }
         buildScriptUpdates = await BuildScriptUpdates.create(
-            _environment.reader, _options.packageGraph, assetGraph);
+            _environment.reader,
+            _options.packageGraph,
+            assetGraph,
+            _options.skipBuildScriptCheck);
         conflictingOutputs = assetGraph.outputs
             .where((n) => n.package == _options.packageGraph.root.name)
             .where(inputSources.contains)