+++ build_runner_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.1.1
+
+- When skipping build script updates, don't check if the build script is a
+  part of the asset graph either.
+
 ## 3.1.0
 
 - Factor out the logic to do a manual file system scan for changes into a